+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add option to view months available as string (e.g., March - November), in addition to the graph.
+- Add option to view times available as string (e.g., 8AM - 7PM), in addition to the graph.
+- Add option to view times as 12-hour (3PM) or 24-hour (15:00) time.
 ### Fixed
 - Fix for off-by-one issue when getting availability, which led to incorrectly showing the months that an item was available.